Pipeline fixes and hiding namespaces count badge
