+++ docs/components/Core.mdx
@@ -231,20 +231,11 @@ The Add Memory component appends one or more items to canvas-level memory storag
 
 ### List Mode
 
-Enable **Input is a list** to write one memory row per element of a list expression.
+Enable "Input is a list" to add one memory row per element of a list expression.
+Configure `listSource` (the expression that yields the list) and `itemVariable`
+(the name to reference each element in field-value expressions, defaults to `item`).
 
-- **List Source**: an expression that evaluates to a list, for example `$["Runner"].data.result.services`.
-- **Item Variable**: the variable name used to reference each element in the value expressions (defaults to `item`).
-
-In list mode, each field value in the **Values** list is evaluated once per element with the iteration variable in scope. A single `memory.added` event is emitted with all written rows and a `count` field.
-
-#### Example Configuration
-
-- List Source: `$["Runner"].data.result.services`
-- Item Variable: `service`
-- Values:
-  - `name` → `service.name`
-  - `id` → `service.id`
+A single `memory.added` event is emitted with all written rows and the total `count`.
 
 ### Example Output
 
@@ -258,25 +249,6 @@ In list mode, each field value in the **Values** list is evaluated once per elem
         "id": "1",
         "pull_request": "123"
       }
-    }
-  },
-  "timestamp": "2026-01-19T12:00:00Z",
-  "type": "memory.added"
-}
-```
-
-#### Example Output (list mode)
-
-```json
-{
-  "data": {
-    "data": {
-      "namespace": "services",
-      "count": 2,
-      "values": [
-        { "name": "api", "id": "1" },
-        { "name": "worker", "id": "2" }
-      ]
     }
   },
   "timestamp": "2026-01-19T12:00:00Z",
@@ -943,12 +915,10 @@ The Update Memory component updates matching rows in canvas-level memory storage
 
 ### List Mode
 
-Enable **Input is a list** to iterate over a list expression and run one update per element.
-
-- **List Source**: an expression that evaluates to a list, for example `$["Runner"].data.result.services`.
-- **Item Variable**: the variable name used to reference each element in the value expressions (defaults to `item`).
-
-The `matchList` is evaluated once and applied identically to every iteration. Each `valueList` field value is evaluated per element with the iteration variable in scope. All resulting updates are reported in a single `memory.updated` event; the channel is `found` when at least one row was updated across all iterations.
+Enable "Input is a list" to iterate over a list expression and run one update per element.
+The `matchList` is evaluated once (same matches for every iteration), while each
+`valueList` field value is evaluated per element with the iteration variable in scope
+(defaults to `item`). All resulting updates are reported in a single `memory.updated` event.
 
 ### Example Output
 
@@ -1012,14 +982,14 @@ Always emits to the default channel. Check `data.operation` to know whether the 
 
 ### List Mode
 
-Enable **Input is a list** to iterate over a list expression and run one upsert per element.
+Enable "Input is a list" to iterate over a list expression and run one upsert per element.
+The `matchList` is evaluated once (same matches for every iteration), while each
+`valueList` field value is evaluated per element with the iteration variable in scope
+(defaults to `item`). All upserts are reported in a single `memory.upserted` event with per-item operation results.
 
-- **List Source**: an expression that evaluates to a list, for example `$["Runner"].data.result.services`.
-- **Item Variable**: the variable name used to reference each element in the value expressions (defaults to `item`).
-
-The `matchList` is evaluated once and applied identically to every iteration. Each `valueList` field value is evaluated per element with the iteration variable in scope. All upserts are reported in a single `memory.upserted` event with a per-item `items` array containing the operation result (`updated` or `created`) and the affected records.
-
-When a non-empty `matchList` is configured in list mode, each iteration attempts to update the same matched rows. List mode is most useful for inserting multiple rows (empty `matchList`) or when each item's values are meant to overwrite the matched rows.
+Note: when a non-empty `matchList` is configured in list mode, each iteration attempts to
+update the same matched rows. List mode is most useful for inserting multiple rows
+(empty `matchList`) or when each item's values are meant to overwrite the matched rows.
 
 ### Example Output
