+++ web_src/src/ui/CanvasPage/components/CanvasModeToggle.spec.tsx
@@ -45,17 +45,15 @@ describe("CanvasModeToggle", () => {
     expect(onSelectMemory).toHaveBeenCalledTimes(1);
   });
 
-  it("renders the namespace count badge when memoryNamespaceCount > 0", () => {
+  it("does not render the namespace count badge while MEMORY_TAB_NAMESPACE_BADGE_ENABLED is false", () => {
     render(
       <CanvasModeToggle mode="version-live" onSelectLive={vi.fn()} onSelectMemory={vi.fn()} memoryNamespaceCount={3} />,
     );
 
-    const badge = screen.getByTestId("canvas-view-mode-memory-badge");
-    expect(badge).toBeInTheDocument();
-    expect(badge).toHaveTextContent("3");
+    expect(screen.queryByTestId("canvas-view-mode-memory-badge")).not.toBeInTheDocument();
   });
 
-  it("hides the badge entirely when memoryNamespaceCount is 0", () => {
+  it("hides the badge when memoryNamespaceCount is 0", () => {
     render(
       <CanvasModeToggle mode="version-live" onSelectLive={vi.fn()} onSelectMemory={vi.fn()} memoryNamespaceCount={0} />,
     );
