+++ web_src/src/ui/CanvasPage/components/CanvasModeToggle.tsx
@@ -23,6 +23,9 @@ const DASHBOARD_TAB = "dashboard";
 const MEMORY_TAB = "memory";
 const RUNS_MODE = "runs";
 
+/** Re-enable when the Memory tab namespace count badge should be visible again. */
+const MEMORY_TAB_NAMESPACE_BADGE_ENABLED = false;
+
 export function CanvasModeToggle({
   mode,
   onSelectLive,
@@ -110,7 +113,7 @@ export function CanvasModeToggle({
           <TabsTrigger value={MEMORY_TAB} data-testid="canvas-view-mode-memory" aria-label="Memory">
             <span className="inline-flex items-center gap-1.5">
               Memory
-              {memoryNamespaceCount > 0 ? (
+              {MEMORY_TAB_NAMESPACE_BADGE_ENABLED && memoryNamespaceCount > 0 ? (
                 <Badge
                   variant="secondary"
                   className="h-4 px-1.5 py-0 text-[10px] leading-none"
